feat: 메세지 시큐리티 설정 재조정
