+++ src/main/java/com/example/auctrade/global/config/MessageSecurityConfig.java
@@ -1,0 +1,38 @@
+package com.example.auctrade.global.config;
+
+import org.springframework.context.annotation.Bean;
+import org.springframework.context.annotation.Configuration;
+import org.springframework.messaging.Message;
+import org.springframework.messaging.support.ChannelInterceptor;
+import org.springframework.security.authorization.AuthorizationManager;
+import org.springframework.security.config.annotation.web.socket.EnableWebSocketSecurity;
+import org.springframework.security.messaging.access.intercept.MessageMatcherDelegatingAuthorizationManager;
+
+import static org.springframework.messaging.simp.SimpMessageType.MESSAGE;
+import static org.springframework.messaging.simp.SimpMessageType.SUBSCRIBE;
+
+@Configuration
+@EnableWebSocketSecurity
+public class MessageSecurityConfig {
+
+    // 웹소켓 관련
+    @Bean
+    public AuthorizationManager<Message<?>> messageAuthorizationManager(MessageMatcherDelegatingAuthorizationManager.Builder messages) {
+        messages
+                .nullDestMatcher().authenticated()
+//                .simpSubscribeDestMatchers("/user/queue/errors").permitAll() // 개인 사용자 큐에 에러 메세지 보낼 때 써먹기
+                .simpDestMatchers("/send/**").hasAnyRole("USER", "ADMIN")
+                .simpSubscribeDestMatchers("/sub/**").hasAnyRole("USER", "ADMIN")
+                .simpTypeMatchers(MESSAGE, SUBSCRIBE).denyAll()
+                .anyMessage().denyAll();
+
+        return messages.build();
+    }
+
+    // csrf 보호 설정 인터셉터 익명 구현 객체
+    @Bean("csrfChannelInterceptor")
+    public ChannelInterceptor csrfChannelInterceptor() {
+        return new ChannelInterceptor() {
+        };
+    }
+}
